+++ name_criterion.py
@@ -1,0 +1,5 @@
+def name_criterion(name):
+    if len(name) > 20:
+        print("User name should be within 20 characters")
+        return name
+    return name
